% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardthemes.R
\docType{package}
\name{dashboardthemes}
\alias{dashboardthemes}
\title{dashboardthemes: a package for custom themes in shinydashboard applications}
\description{
The dashboardthemes package provides two main important features:
\itemize{
  \item Using new pre-defined themes and logos for dashboards.
  \item Creating custom themes and logos for dashboards.
}
}
\section{Available functions}{

The available functions are:
\itemize{
  \item \code{\link{shinyDashboardThemes}}: Calls a custom theme created using shinyDashboardThemeDIY.
  \item \code{\link{shinyDashboardLogo}}: Calls a custom logo created using shinyDashboardLogoDIY.
  \item \code{\link{shinyDashboardThemeDIY}}: Creates a custom theme object.
  \item \code{\link{shinyDashboardLogoDIY}}: Creates a custom logo object. Inspired from \href{http://www.dataseries.org/}{here}.
  \item \code{\link{cssGradientThreeColors}}: Creates a three colour gradient to be used in themes.
}
}

