% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_BlueGradient.R
\name{logo_blue_gradient}
\alias{logo_blue_gradient}
\title{logo_blue_gradient}
\usage{
logo_blue_gradient(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\arguments{
\item{boldText}{String. Bold text for the logo.}

\item{mainText}{String. Main text for the logo.}

\item{badgeText}{String. Text for the logo badge.}
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Blue Gradient logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
