% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_FlatRed.R
\name{logo_flat_red}
\alias{logo_flat_red}
\title{logo_flat_red}
\usage{
logo_flat_red(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Flat Red logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
