% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_GreyDark.R
\name{logo_grey_dark}
\alias{logo_grey_dark}
\title{logo_grey_dark}
\usage{
logo_grey_dark(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Grey Dark logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
