% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_GreyLight.R
\name{logo_grey_light}
\alias{logo_grey_light}
\title{logo_grey_light}
\usage{
logo_grey_light(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\arguments{
\item{boldText}{String. Bold text for the logo.}

\item{mainText}{String. Main text for the logo.}

\item{badgeText}{String. Text for the logo badge.}
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Grey Light logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
