% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_PoorMansFlatly.R
\name{logo_poor_mans_flatly}
\alias{logo_poor_mans_flatly}
\title{logo_poor_mans_flatly}
\usage{
logo_poor_mans_flatly(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Poor Man's Flatly logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
