% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_PurpleGradient.R
\name{logo_purple_gradient}
\alias{logo_purple_gradient}
\title{logo_purple_gradient}
\usage{
logo_purple_gradient(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
Purple Gradient logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
