% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_BlueGradient.R
\docType{data}
\name{theme_blue_gradient}
\alias{theme_blue_gradient}
\title{theme_blue_gradient}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_blue_gradient
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Blue Gradient theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
