% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_FlatRed.R
\docType{data}
\name{theme_flat_red}
\alias{theme_flat_red}
\title{theme_flat_red}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_flat_red
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Flat Red theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
