% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_GreyDark.R
\docType{data}
\name{theme_grey_dark}
\alias{theme_grey_dark}
\title{theme_grey_dark}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_grey_dark
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Grey Dark theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
