% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_GreyLight.R
\docType{data}
\name{theme_grey_light}
\alias{theme_grey_light}
\title{theme_grey_light}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_grey_light
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Grey Light theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
