% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_OneNote.R
\docType{data}
\name{theme_onenote}
\alias{theme_onenote}
\title{theme_onenote}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_onenote
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
OneNote theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
