% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_PoorMansFlatly.R
\docType{data}
\name{theme_poor_mans_flatly}
\alias{theme_poor_mans_flatly}
\title{theme_poor_mans_flatly}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_poor_mans_flatly
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Poor Man's Flatly theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
