% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_PurpleGradient.R
\docType{data}
\name{theme_purple_gradient}
\alias{theme_purple_gradient}
\title{theme_purple_gradient}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_purple_gradient
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Purple Gradient theme for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
