% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_OneNote.R
\name{logo_onenote}
\alias{logo_onenote}
\title{logo_onenote}
\usage{
logo_onenote(boldText = "Shiny", mainText = "App", badgeText = "v1.1")
}
\arguments{
\item{boldText}{String. Bold text for the logo.}

\item{mainText}{String. Main text for the logo.}

\item{badgeText}{String. Text for the logo badge.}
}
\value{
Object produced by shinyDashboardLogoDIY
}
\description{
OneNote logo for a shinydashboard application
}
\seealso{
\code{\link{shinyDashboardLogoDIY}}
}
